// libs/api/produttori-service/src/lib/model/produttore-create-dto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProduttoreCreateDTO { 
    nome: string;
    cognome: string;
    codiceFiscale: string;
    numeroPrivato: string;
    email: string;
    partitaIva: string;
    tesseramentoAttivo?: boolean;
}